partial revert: rename module: math -> sse_math, to avoid name conflicts
