+++ gorgon/toolkit/sse/hunter.py
@@ -1,6 +1,6 @@
 import logging
 from ..libpytoolkit import *
-from .math import *
+from .sse_math import *
 from termcolor import colored
